move jsonkey
